d2: Move cmd to repo root
Closes #113
