+++ ci/release/_build.sh
@@ -14,7 +14,7 @@ export GOOS=$(goos "$OS")
 export GOARCH="$ARCH"
 sh_c mkdir -p "$HW_BUILD_DIR/bin"
 sh_c go build -ldflags "'-X oss.terrastruct.com/d2/lib/version.Version=$VERSION'" \
-  -o "$HW_BUILD_DIR/bin/d2" ./cmd/d2
+  -o "$HW_BUILD_DIR/bin/d2" .
 
 ARCHIVE=$PWD/$ARCHIVE
 cd "$(dirname "$HW_BUILD_DIR")"
