+++ d2lib/c.go
@@ -1,6 +1,6 @@
 //go:build cgo
 
-package d2
+package d2lib
 
 import "oss.terrastruct.com/d2/d2layouts/d2dagrelayout"
